Refactor: 코드 포매팅 (#22)
