+++ src/main/java/com/flytrap/rssreader/domain/member/Member.java
@@ -2,11 +2,9 @@ package com.flytrap.rssreader.domain.member;
 
 import com.flytrap.rssreader.global.model.DefaultDomain;
 import com.flytrap.rssreader.global.model.Domain;
-import com.flytrap.rssreader.infrastructure.entity.member.OauthServer;
 import java.io.Serializable;
 import java.time.Instant;
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -24,8 +22,9 @@ public class Member extends DefaultDomain implements Serializable {
     private Instant createdAt;
 
     @Builder
-    private Member(Long id, String name, String email, String profile, long oauthPk, OauthServer oauthServer,
-        Instant createdAt) {
+    private Member(Long id, String name, String email, String profile, long oauthPk,
+            OauthServer oauthServer,
+            Instant createdAt) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -34,16 +33,17 @@ public class Member extends DefaultDomain implements Serializable {
         this.createdAt = createdAt;
     }
 
-    public static Member of(Long id, String name, String email, String profile, Long oauthPk, OauthServer oauthServer, Instant createdAt) {
+    public static Member of(Long id, String name, String email, String profile, Long oauthPk,
+            OauthServer oauthServer, Instant createdAt) {
         return Member.builder()
-            .id(id)
-            .name(name)
-            .email(email)
-            .profile(profile)
-            .oauthPk(oauthPk)
-            .oauthServer(oauthServer)
-            .createdAt(createdAt)
-            .build();
+                .id(id)
+                .name(name)
+                .email(email)
+                .profile(profile)
+                .oauthPk(oauthPk)
+                .oauthServer(oauthServer)
+                .createdAt(createdAt)
+                .build();
     }
 
     public Long getOauthPk() {
@@ -53,11 +53,4 @@ public class Member extends DefaultDomain implements Serializable {
     public OauthServer getOauthServer() {
         return this.oauthInfo.getOauthServer();
     }
-}
-
-@Getter
-@AllArgsConstructor
-class OauthInfo implements Serializable {
-    private Long oauthPk;
-    private OauthServer oauthServer;
 }
