+++ src/test/java/com/flytrap/rssreader/service/AuthServiceTest.java
@@ -1,12 +1,16 @@
 package com.flytrap.rssreader.service;
 
-import static org.mockito.BDDMockito.*;
+import static org.mockito.BDDMockito.any;
+import static org.mockito.BDDMockito.anyString;
+import static org.mockito.BDDMockito.times;
+import static org.mockito.BDDMockito.verify;
+import static org.mockito.BDDMockito.when;
 
 import com.flytrap.rssreader.domain.member.Member;
+import com.flytrap.rssreader.domain.member.OauthServer;
 import com.flytrap.rssreader.infrastructure.api.AuthProvider;
 import com.flytrap.rssreader.infrastructure.api.dto.AccessToken;
 import com.flytrap.rssreader.infrastructure.api.dto.UserResource;
-import com.flytrap.rssreader.infrastructure.entity.member.OauthServer;
 import com.flytrap.rssreader.presentation.dto.Login;
 import jakarta.servlet.http.HttpSession;
 import org.assertj.core.api.SoftAssertions;
@@ -38,20 +42,22 @@ class AuthServiceTest {
     @BeforeEach
     void init() {
         when(authProvider.requestAccessToken(anyString()))
-            .thenReturn(Mono.just(new AccessToken("test_access_token", "Bearer")));
+                .thenReturn(Mono.just(new AccessToken("test_access_token", "Bearer")));
         when(authProvider.requestUserResource(any()))
-            .thenReturn(Mono.just(UserResource.builder().id(1L).email("test@gmail.com").login("login").avatarUrl("img.jpg").build()));
+                .thenReturn(Mono.just(
+                        UserResource.builder().id(1L).email("test@gmail.com").login("login")
+                                .avatarUrl("img.jpg").build()));
         when(memberService.loginMember(any()))
-            .thenReturn(
-                Member.builder()
-                    .id(1L)
-                    .name("테스트")
-                    .email("test@gmail.com")
-                    .profile("img.jpg")
-                    .oauthPk(11L)
-                    .oauthServer(OauthServer.GITHUB)
-                    .build()
-            );
+                .thenReturn(
+                        Member.builder()
+                                .id(1L)
+                                .name("테스트")
+                                .email("test@gmail.com")
+                                .profile("img.jpg")
+                                .oauthPk(11L)
+                                .oauthServer(OauthServer.GITHUB)
+                                .build()
+                );
     }
 
     @Test
